refactor(socket): shrink SocketClient to thin transport adapter
Extract SocketCommandContextImpl and MonsterLineFormatter into their own
classes; move command-context construction into SocketCommandDispatcher;
add ArchUnit guard to prevent SocketClient regrowth beyond transport concerns.

Closes #182
