+++ src/test/java/io/taanielo/jmud/core/server/socket/MonstersLineFormatterTest.java
@@ -11,7 +11,7 @@ import io.taanielo.jmud.core.mob.MobTemplate;
 import io.taanielo.jmud.core.world.RoomId;
 
 /**
- * Unit tests for {@link SocketClient#formatMonstersLine(List)}.
+ * Unit tests for {@link MonsterLineFormatter#format(List)}.
  */
 class MonstersLineFormatterTest {
 
@@ -36,35 +36,35 @@ class MonstersLineFormatterTest {
 
     @Test
     void emptyListReturnsNone() {
-        assertEquals("Monsters: none", SocketClient.formatMonstersLine(List.of()));
+        assertEquals("Monsters: none", MonsterLineFormatter.format(List.of()));
     }
 
     @Test
     void nullListReturnsNone() {
-        assertEquals("Monsters: none", SocketClient.formatMonstersLine(null));
+        assertEquals("Monsters: none", MonsterLineFormatter.format(null));
     }
 
     @Test
     void singleMobShowsOneEntry() {
-        assertEquals("Monsters: 1x Goblin", SocketClient.formatMonstersLine(List.of(mob("Goblin"))));
+        assertEquals("Monsters: 1x Goblin", MonsterLineFormatter.format(List.of(mob("Goblin"))));
     }
 
     @Test
     void multipleDifferentMobsListedSeparately() {
-        String result = SocketClient.formatMonstersLine(List.of(mob("Goblin"), mob("Troll")));
+        String result = MonsterLineFormatter.format(List.of(mob("Goblin"), mob("Troll")));
         assertEquals("Monsters: 1x Goblin, 1x Troll", result);
     }
 
     @Test
     void sameTypeMobsAreGroupedAndCounted() {
-        String result = SocketClient.formatMonstersLine(
+        String result = MonsterLineFormatter.format(
             List.of(mob("Goblin"), mob("Goblin"), mob("Goblin")));
         assertEquals("Monsters: 3x Goblin", result);
     }
 
     @Test
     void mixedMobsGroupedByName() {
-        String result = SocketClient.formatMonstersLine(
+        String result = MonsterLineFormatter.format(
             List.of(mob("Goblin"), mob("Troll"), mob("Goblin")));
         assertEquals("Monsters: 2x Goblin, 1x Troll", result);
     }
@@ -72,7 +72,7 @@ class MonstersLineFormatterTest {
     @Test
     void encounterOrderPreservedForFirstOccurrence() {
         // Troll appears first, then Goblin — output must match insertion order.
-        String result = SocketClient.formatMonstersLine(
+        String result = MonsterLineFormatter.format(
             List.of(mob("Troll"), mob("Goblin"), mob("Troll")));
         assertEquals("Monsters: 2x Troll, 1x Goblin", result);
     }
